^R Extract host-state/mounts/audit cluster into internal/host/hoststate
Fifth sub-PR of the hostutil split. Moves the mount-cache key,
host-output path resolution, audit-chain digest, manifest/resolver
metadata readers, and stale-resource cleanup helpers from
internal/hostutil/launcher.go into a new internal/host/hoststate/
package. Drops ~430 LOC from internal/hostutil/launcher.go.

Moved exports: AuditRecordDigest, DirectMountCacheKey,
WorkspaceCacheKey, ResolveHostOutputCandidate,
ResolveHostOutputDirectoryCandidate, ManifestMetadataLines,
ResolverMetadataLines, CleanupStaleLatestLogPointers,
CleanupStaleSessionAuditDirs, CleanupStaleInjectionBundles.

Also moved: the private injectionBundleIsLive plus the shared
unexported helpers (renderStringMap/stringSlice/stringOrEmpty/
stringOrDefault/boolTo01/readJSONFile/currentUID/statUID/
expandUserPathForLauncher) since none are needed by what stays
in hostutil. The dead isSymlink helper is deleted.

cmd/workcell-hostutil/main.go and the two external test files
(authresolve, cmd/workcell-hostutil main_test) update their
imports accordingly.

go build/test/gofmt clean; no behavior change.
