+++ internal/host/hoststate/doc.go
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Omkhar Arasaratnam
+
+// Package hoststate provides host-side state and resource accounting
+// helpers used by the launcher control plane: mount/workspace cache
+// key derivation, host-output path resolution, audit-chain digesting,
+// manifest/resolver metadata extraction, and stale-resource cleanup
+// for log pointers, session audit dirs, and injection bundles.
+package hoststate
